changing the display_numeric_data abd the display_string to use the h5dump_sprint from the tools lib.
added a section to initialize the h5dump_t variable in both these functions.
in the display_string had to remove the extra quotes that gets displayed when
printing.

in h5dump_sprint i changed the for loop in the case of the string.
I added a check to see if the last character is a \0.  it was being printed
before.