Purpose: Improvement
Description:
The cvs checkout used the complete path of current/ which is not
accepted by cvs server.  Changed it to chdir to the snapshot directory
and then cvs checkout just current/.

Platforms tested:
Not h5committested because it does not test this command.
Tested in tg-ncsa which has no direct access to the CVS directory
and must use the server route.

Misc. update: