Purpose:  Bug fix
Description:  A new bug is found on HP.  There is float exception during conversion from double to
unsigned long long when the value of double is very big.

Solution:  Try to catch the problem in configure and skip this part of test.

Platforms tested:  kelgia and verbena(mainly these two machines are involved)