Purpose: Small H5I API changes.
Description:
Changed public version of H5Inmembers to return the number of IDs in a
type by reference.  This allows the user to differentiate between the
case when H5Inmembers has an error and when the ID type is just empty.
Added a new API function, H5Itype_exists, which allows the user to
find out whether an ID type exists or not, without throwing an error.

Solution:

Platforms tested:
   Copper

Misc. update: