Purpose:  Correction to previous bug fix.  This checkin is mainly to let user test his program.
Description:   For family driver bug(#213), the previous fix was imperfect.  The problem
was when user create family file, the member size information wasn't saved in file.  When
the file was re-opened, the library simply use the size of 1st member file as member size.

Solution:
              When file is re-opened, member size passed in from access property
              is checked to see if it's reasonable.  If there is only 1 member
              file, member size can't be smaller than current member size.
              If there are at least 2 member files, member size can only be equal
              to the 1st member size.

Platforms tested:  h5committest and fuss