Purpose: Bug fix
Description:
Discovered two small bugs.
When MAKE is defined as the empty string, a test in configure incorrectly
identifies it as pmake.  Skipped this test if $MAKE is the empty string.
Two .h5 files produced by a test in the hl directory were not cleaned.
Updated the Makefile.am.


Platforms tested:
heping, modi4 (very minor changes)