Purpose: bug fix
Description:
an assertion after the search for objects was saying that the number of
objects in file should be > 0
removed it, since there can be a case where the number of objects is 0.
in this case this was found by a case where zlib was not present and the output file
was not written any objects to

Solution:

Platforms tested:
linux

Misc. update: