Purpose: Improvement
Description:
The previous HTML files strategy was to expect the very first comment
block is the copyright notice. That did not work well as more and more
other comments appear now.

Solution:
Changed the strategy to look for the only the essential body of copyright
notice.  This no longer verifies this body of text is a comment block. But
the previous strategy is not really bullet-proof either.

Platforms tested:
tested in heping.