Purpose: bug fix.
Description:
Fixed the segmentation fault errors in modi4, copper and tg-login.
It was due to the misuse of trying to realloc a pointer returned by
getenv_all. (not supposed to.)

Also rearranged the code so that option is checked first, then check
with environment variable, then use default setup.  This saves the
need to do realloc at all.

Platforms tested:
Heping, modi4, shanti, copper (copper showed a different error now.)