Purpose: Moving gif2h5 to hl directory in 1.6 branch
Description:
This is the same change as was made in the 1.7 branch.

Solution:
Moved the gif2h5 tool from the hdf5/tools directory to the hdf5/hl/tools
directory.  The files were already present, but now the Makefiles
build and install using the new directory instead of the old one.

Platforms tested:
mir, modi4