Fix (hopefully) for last night's failures on tg-login3.
It seems that the latest version of libtool is linking against shared versions
of the szip libraries in Fortran, rather than the static versions.  ifort
doesn't support shared libraries at all.

To solve the problem, we avoid shared libraries entirely unless we're
building shared Fortran libraries ourselves.

If libtool always behaves like this now, it may cause similar problems on
other machines.  If you notice that a machine is breaking when it links against
a shared library, let me know.

Tested on kagiso; nothing new is broken.  If Albert has time this afternoon
I'll ask him to test on tg-login3 as well.