Purpose: Bug fix.
Description:
The fortran configure tried to test link with the MPI library but it
used "mpi_file_open()" which is actually an MPI-IO API.  It might not
always be included in the -lmpi library.  It looked like an error when
the configure code was adapted from the top configure.in.
Why it worked before: it was probably because the top configure has
already figured out -lmpio is needed and this information is sent to
the fortran configure in older version of autoconf.  The latest one
does not do that and exposed this error in LANL QSC.

Solution:
Fixed it to use the proper "mpi_init()" instead.

Tested Platform:
LANL QSC and kagiso, both using parallel.