Purpose: Update test directories in MinGW documentation
Description:
There are many tests that are unsupported on MinGW because of the way it handles certain syntax in shell scripts (particularly '/').  As a result, we list a few tests that users must comment out in various Makefile.in's.  This checkin updates the location of h5stat-- it was formerly in ./tools/misc, but has been moved it ./tools/h5stat.

Tested:
MinGW on WinXP