Purpose: Work around Windows FC bug in h5dump xml test script
Description:
In some rare instances, FC (diff equivalent) command on Windows will fail when comparing ASCII files with opposite line-ending formats (Unix uses LF, while Windows uses CRLF).  This was an issue for one test case in our test script.  It only came up when testing from a release tarball, because the expected output was not converted to Windows-style.  In this situation, FC was reporting false differences.

To workaround, we double-check failure cases and convert the EOL characters if neccessary.

Tested:
VS2005 on WinXP