Purpose: Check for __func__ keyword during configure
Description:
In some of the new H5EA* code, many new C99 features are being used.  Most of them are portable to all compilers, but the __func__ keyword in particular isn't supported in Visual Studio on Windows.  Instead, Windows defines the __FUNCTION__ keyword, which can be used as a direct substitute.

We now check for the __func__ keyword during configure and define the feature flag H5_HAVE_C99_FUNC.  There was previously a check for __FUNCTION__, and the feature flags H5_HAVE_FUNCTION.  In H5EApkg.h, we check for the presents of each of these in order to see which to use.  If neither are avaiable, fail.

Tested:
h5committest (kagiso, smirom, linew)