Purpose: Declare loop variable at beginning of function
Purpose:
On Windows, we were getting compile errors because h5test.c includes a function with variable declarations in the middle of the call.  The Microsoft C compiler demands that all variables be declared at the top of the function.  This checkin simply moves the declaration to the top to satisfy Visual Studio.

Tested:
VS2005 on WinXP