Bug fix: 1917.
Description:
test/big incorrectly determined not able to write files larger than 2GB and
skipped the SEC2 and STDIO driver tests.  The reason was because it was using
off_t while the SEC2 driver is using lseek64 which expects off64_t type.

Solution:
Created a new HDoff_t which is set to off_t or off64_t or other appropriate
type depending on which of lseek or lseek64 is available. Changed SEC2 file
driver and the big test to use this common definition.

Tested:
In BP (AIX), using --enable and --disable-largefile, for both 32 and 64 bits
modes.  Did not do h5committest because: 1. the error was exposed in the remote
BP machine; 2. the change is trivial.