Purpose: Add enhancements to direct I/O driver
Description:
Adds support to chunk cache to allocate memory that can be used directly with
the direct I/O driver.  This will work when the file alignment is set properly
and reading or writing unfiltered datasets, or when reading filtered datasets.
When writing filtered datasets, the driver will still have to reallocate the
buffer (unless the filters didn't reallocate).

Also added public functions H5Pset_aligned_mem() and H5Pget_aligned_mem() which
can be used to signify that the application allocated a buffer in a manner that
can be used with the direct I/O driver.

Tested: fedora, ember