Description:
Changed the length of the fortran string passed to HD5packFstring in the function nh5pget_external_c. The 3rd
argument should be the fortran length of the string, not the C length of the string (which includes the null).

Tested: jam (intel and gnu), also checked the example h5ex_d_extern.f90 which detected the problem on Amazon c2
machine (reported by Larry).