Description:
Removed line:

$(RM) $(DESTDIR)$(includedir)/H5f90i*.h

These *.h files are needed for HL_NPOESS and should not be removed when make clean is specified. 

The Daily Test installs the file using the deploy script, but then
runs make clean which removes the file, so the files are not there when we try to compile HL_NPOESS
causing an error.