Purpose: Fix testflushrefresh failures
Description:
Since the "flush me last" code was introduced, it was no longer possible for the
superblock to be flushed to disk without flushing the entire file.  This made it
impossible to open a file concurrently, even after calling H5Gflush on the root
group.  The superblock was originally supposed to be flushed every time any
object was explicitly flushed with H5*flush.

Modified H5C_flush_cache to ignore flush me last flag if the entry is marked for
flush (which is done either by a tag for H5*flush or to free up space in cache)
and the flush marked entries flag is set.

Tested: ummon