Merge r23366 from trunk.  Fixes HDFFV-8296.
Removed the H5Pset_dxpl_multi and H5Pget_dxpl_multi functions from
the library.  The intended functionality for them was never fully
implemented and they are fundamentally broken.

The functions were removed from the C and C++ interfaces.  They were
not exported in the Fortran interface.

Tested on jam