Bug fix: Mac system has dynamic library name in the form of libxyz.dylib. Solution: Changed test_plugin.in to copy library file names in the form of libdynlib{123].*.
Another problem:
Even when copy failed in copying the libray files (was in the form of *.so*),
the test still passed.  That was because .libs was included in $HDF5_PLUGIN_PATH and the needed
plug in libraries were generated there.
Solution:
Fixed by copying the plug in library files to separate directories and seetup HDF5_PLUGIN_PATH
to include them but NOT .libs.

Tested: used desycommittest (duck, emu, jam, koala, ostrich) plus cmakehdf5 in jam.