Merge of r26071 from features/autotools_rework
Removes the --with-default-vfd option from configure. This was basically
useless and only allowed selecting the stdio VFD. Since this is a demo
VFD and not really a production VFD, we decided to retire this option.

Fixes: HDFFV-9081

Tested on: h5committest

NOTE: platypus cmake fails, but this is not a new error.