Interim check in of the cache image feature.
Serial version is fully implemented with the exception of the 
code to save and reload adaptive cache resize status.

Parallel version is at least partially implements -- need to 
review flush process to see if serial code to supress entry 
writes is sufficient in the parallel case.

Passes all existing tests in the serial and parallel case.
Passes moderate smoke checks in the serial case.  No new 
tests yet for the parallel case.

Tested serial and parallel on Mercury and Jam.